backend/requirements.txt
Flask==2.2.5
tensorflow==2.20.0
Pillow==12.0.0